Positive anything is better than negative nothing

+++ 0x01-variables_if_else_while/0-positive_or_negative.c
@@ -6,18 +6,18 @@
  */
 int main(void)
 {
-long int n = rand();
+int n = rand();
 if (n > 0)
 {
-printf("%li is positive\n", n);
+printf("%i is positive\n", n);
 }
 else if (n == 0)
 {
-printf("%li is zero\n", n);
+printf("%i is zero\n", n);
 }
 else
 {
-printf("%li is negative\n", n);
+printf("%i is negative\n", n);
 }
 return (0);
 }
